pyproject.toml
[project]
name = "siamese-protein-structures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "dataset>=1.6.2",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
